+++ .github/workflows/ReleaseNotes.md
@@ -1,4 +1,6 @@
 * [Client] Fixed _PlatformNotSupportedException_ when using Blazor (#1755, thanks to @Nickztar).
+* [Client] Added hot reload of client certificates (#1781).
+* [Client] Added several new option builders and aligned usage (#1781, BREAKING CHANGE!).
 * [Client] Added support for _RemoteCertificateValidationCallback_ for .NET 4.5.2, 4.6.1 and 4.8 (#1806, thanks to @troky).
 * [Client] Fixed wrong logging of obsolete feature when connection was not successful (#1801, thanks to @ramonsmits).
 * [Client] Fixed _NullReferenceException_ when performing several actions when not connected (#1800, thanks to @ramonsmits).